handle rename events

+++ Rakefile
@@ -13,12 +13,23 @@ namespace :authors do
     puts issues.size.to_s + ' issues'
     puts events.size.to_s + ' events'
 
+    renamed_events = events.filter{|e| e.event == 'renamed' && e.created_at > last_update_time}
+    unless renamed_events.empty?
+      renamed_events.each do |e| 
+        author = Author.where(github_issue_num: e.issue.number).first
+        print "rename author #{author.name} to #{e.issue.title} ... "  
+        author.update(name: e.issue.title)
+        puts 'done'
+      end
+    end
+
     issues.each do |issue|
       puts 'issue ' + issue.number.to_s
       authors_found = Author.where(github_issue_num: issue.number)
       if authors_found.empty?
-        puts "create author #{issue.title} - #{issue.number}"
+        print "create author #{issue.title} - #{issue.number} ... "
         Author.create(name: issue.title, bio: issue.body, github_issue_num: issue.number)
+        puts 'done'
       end
       
       comments = issues_api.comments.list(
@@ -28,16 +39,20 @@ namespace :authors do
         since: last_update_time
       )
       unless comments.empty?
-        puts "update author #{issue.title} - #{issue.number}"
+        print "update bio for author #{issue.title} - #{issue.number} ... "
         last_comment = comments.to_a.last.body
         authors_found.first.update(bio: last_comment)
+        puts 'done'
       end
     end
 
-    closed_events = events.filter{|e| e.event == 'closed'}
-    closed_github_issue_nums = closed_events.map{|e| e.issue.number}
-    Author.where(github_issue_num: closed_github_issue_nums).destroy_all
-    closed_events.each{|e| puts "delete author #{e.issue.title}"}
+    closed_events = events.filter{|e| e.event == 'closed' && e.created_at > last_update_time}
+    unless closed_events.empty?
+      closed_github_issue_nums = closed_events.map{|e| e.issue.number}
+      closed_events.each{|e| puts "delete author #{e.issue.title}"}
+      Author.where(github_issue_num: closed_github_issue_nums).destroy_all
+      puts 'done deleting authors'
+    end
 
     new_timestamp = 0.days.ago.iso8601
     Setting.where(name: 'github_update').first.update(value: new_timestamp)
